CYB-2: remove 1MB file, add gemspec file, for build: gem build cdap-stream-client.gemspec, for test: ruby tests/tests.rb
