+++ .gitignore
@@ -27,3 +27,5 @@ logs/
 # Examples Stuff
 dependency-reduced-pom.xml
 
+# Ruby
+*.gem
