+++ stream-client/stream-client-ruby/lib/cdap-stream-client.rb
@@ -1,0 +1,1 @@
+require './lib/cdap-stream-client/stream_client'
