+++ stream-client/stream-client-ruby/lib/cdap-stream-client/rest.rb
@@ -1,6 +1,6 @@
 require 'httparty'
 
-class StreamClient
+module CDAP
 
   class Rest
     include HTTParty
