--- stream-client/stream-client-ruby/test/test.rb
@@ -1,3 +1,0 @@
-require 'test/unit'
-require './test_stream_client'
-require './test_stream_writer'
