+++ stream-client/stream-client-ruby/tests/tests.rb
@@ -1,0 +1,3 @@
+require 'test/unit'
+require './tests/test_stream_client'
+require './tests/test_stream_writer'
